# backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_document_api_documents_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_upload_post",
        "type": "object"
      },
      "ChangePasswordRequest": {
        "description": "Payload for ``POST /api/auth/change-password``.",
        "properties": {
          "current_password": {
            "format": "password",
            "title": "Current Password",
            "type": "string",
            "writeOnly": true
          },
          "new_password": {
            "format": "password",
            "minLength": 4,
            "title": "New Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
      "ChatMessage": {
        "description": "One transcript entry exchanged between the user and the assistant.",
        "properties": {
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "role",
          "text"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatSessionDetail": {
        "description": "Full chat session document including the message transcript.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "title",
          "messages",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSessionDetail",
        "type": "object"
      },
      "ChatSessionSummary": {
        "description": "One entry in the chat session list (metadata only, no messages).",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSessionSummary",
        "type": "object"
      },
      "DocumentSummary": {
        "description": "One entry in the uploaded-document listing.",
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "embedding_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding Method"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "workflow_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Name"
          }
        },
        "required": [
          "filename",
          "chunk_count"
        ],
        "title": "DocumentSummary",
        "type": "object"
      },
      "ForceLogoutResponse": {
        "description": "Returned after bumping the global JWT version (invalidates all tokens).",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "title": "ForceLogoutResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobStatusResponse": {
        "description": "Current embedding-generation state for a previously uploaded filename.",
        "properties": {
          "chunk_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Count"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "processed_files": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed Files"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_files": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Files"
          }
        },
        "required": [
          "status"
        ],
        "title": "JobStatusResponse",
        "type": "object"
      },
      "LoginRequest": {
        "description": "Credentials submitted to ``POST /api/auth/login``.",
        "properties": {
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "description": "Issued JWT plus the flag that drives the password-change flow.",
        "properties": {
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "token",
          "username",
          "must_change_password"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "MeResponse": {
        "description": "Profile summary for the caller's JWT.",
        "properties": {
          "must_change_password": {
            "title": "Must Change Password",
            "type": "boolean"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "must_change_password"
        ],
        "title": "MeResponse",
        "type": "object"
      },
      "OkMessageResponse": {
        "description": "Acknowledgement with an optional human-readable message.",
        "example": {
          "ok": true
        },
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "title": "OkMessageResponse",
        "type": "object"
      },
      "OkResponse": {
        "description": "Generic acknowledgement payload used by idempotent mutating endpoints.",
        "example": {
          "ok": true
        },
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
      "SaveSessionRequest": {
        "description": "Payload for persisting a chat session's title and message history.",
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "messages"
        ],
        "title": "SaveSessionRequest",
        "type": "object"
      },
...
      "UploadResponse": {
        "description": "Returned immediately after a PDF upload is accepted for processing.",
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "status": {
            "default": "vectorizing",
            "description": "Lifecycle state: vectorizing, completed, failed, unknown.",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "chunk_count"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Couchbase Realtime Voice RAG",
      "url": "https://github.com/"
    },
    "description": "Reference backend for the Couchbase Realtime Voice RAG demo.\n\nThe API exposes authentication, document ingestion, chat-history persistence,\nCouchbase connection settings, and the WebSocket relay that bridges browser\naudio to the Azure OpenAI Realtime API. Document uploads are chunked and\neither embedded locally via ``text-embedding-3-small`` or vectorized through\nthe Capella AI Services workflow before being indexed for vector search.\n",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Couchbase Realtime Voice RAG",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/change-password": {
      "post": {
        "description": "Rotate the password for the authenticated user and invalidate outstanding JWTs.",
        "operationId": "change_password_api_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change the current user's password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/force-logout": {
      "post": {
        "description": "Bump the global token version so every previously issued JWT is rejected.",
        "operationId": "force_logout_api_auth_force_logout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForceLogoutResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invalidate every issued JWT",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Verify credentials and return a 24-hour JWT in both the body and an HTTP-only cookie.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authenticate and issue a JWT",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/me": {
      "get": {
        "description": "Return the username embedded in the caller's JWT plus any forced-rotation flag.",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Return the caller's profile",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/chat/sessions": {
      "get": {
        "description": "Return every persisted chat session ordered by ``updated_at`` descending.",
        "operationId": "list_sessions_api_chat_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatSessionSummary"
                  },
                  "title": "Response List Sessions Api Chat Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List saved chat sessions",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/sessions/{session_id}": {
      "delete": {
        "description": "Remove the persisted chat session document identified by ``session_id``.",
        "operationId": "delete_session_api_chat_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a chat session",
        "tags": [
          "chat"
        ]
      },
      "get": {
        "description": "Return the stored title, timestamps, and full message transcript for ``session_id``.",
        "operationId": "load_session_api_chat_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Load a chat session",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "description": "Create or overwrite the persisted transcript for ``session_id``.",
        "operationId": "save_session_api_chat_sessions__session_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert a chat session",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/documents": {
      "get": {
        "description": "Aggregate stored chunks by source filename and return per-document summaries.",
        "operationId": "list_documents_api_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummary"
                  },
                  "title": "Response List Documents Api Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List uploaded documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/status/{filename}": {
      "get": {
        "description": "Return the lifecycle state of the background embedding job for ``filename``.",
        "operationId": "get_upload_status_api_documents_status__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Check embedding-generation status",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/upload": {
      "post": {
        "description": "Validate, store, and schedule embedding generation for an uploaded document.",
        "operationId": "upload_document_api_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload a PDF/DOCX/TXT document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/documents/{filename}": {
      "delete": {
        "description": "Remove every stored chunk whose ``source_filename`` matches ``filename``.",
        "operationId": "delete_document_api_documents__filename__delete",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete a document and its chunks",
        "tags": [
          "documents"
        ]
      }
    },
...
  },
...
}